requirements.txt
alembic==1.13.1
bcrypt==4.1.2
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
cuid2==2.0.0
elastic-transport==8.12.0
elasticsearch==8.12.1
Flask==3.0.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pluggy==1.4.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.27
tdqm==0.0.1
tqdm==4.66.2
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
